subtitucion de graphic x dashboard
